meta-qti-data: Fix module auto loading for datarmnet
Datarmnet modules were not able to auto load via systemd
due to a couple things. First, the service needed to be symlinked to
a specific target, second, the script needed to have a correct path to
the on device rootfs location where the dlkms are housed, and
finally, the ExecStop command was removed from the rmnetcore.service
file.

Change-Id: I82e3a67204f817eb35f95653ebdefc543f4b3419
Signed-off-by: Conner Huff <chuff@codeaurora.org>